+++ packages/neon-js/src/experimental/index.ts
@@ -1,6 +1,5 @@
 import * as nep5 from "./nep5";
 import { SmartContract } from "./contract";
-import { wallet } from "@cityofzion/neon-core";
 import {
   getSystemFee,
   calculateNetworkFee,
@@ -15,13 +14,5 @@ const txHelpers = {
   setBlockExpiry,
   addFees,
 };
-
-export interface CommonConfig {
-  networkMagic: number;
-  rpcAddress: string;
-  prioritisationFee?: number;
-  blocksTillExpiry?: number;
-  account?: wallet.Account;
-}
 
 export { nep5, txHelpers, deployContract, SmartContract };
