+++ packages/neon-js/src/experimental/types.ts
@@ -1,0 +1,8 @@
+import { wallet } from "@cityofzion/neon-core";
+export interface CommonConfig {
+  networkMagic: number;
+  rpcAddress: string;
+  prioritisationFee?: number;
+  blocksTillExpiry?: number;
+  account?: wallet.Account;
+}
